\name{mmds-package}
\alias{mmds-package}
\alias{mmds}
\docType{package}
\title{
\packageTitle{mmds}
}
\description{
\packageDescription{mmds}
}

\author{
\packageAuthor{mmds}

Maintainer: \packageMaintainer{mmds}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
\code{\link{MMDS}}, \code{\link{MMDS.cpp}}
}
\examples{
MMDS(X = matrix(c(2,3,5,2,3,7,8,4,5), 3, 3))
MMDS.cpp(X = matrix(c(2,3,5,2,3,7,8,4,5), 3, 3))
}
