\name{mmds-package}
\alias{mmds-package}
\alias{mmds}
\docType{package}
\title{
\packageTitle{mmds}
}
\description{
\packageDescription{mmds}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mmds}
\packageIndices{mmds}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mmds}

Maintainer: \packageMaintainer{mmds}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
MMDS.cpp(X=matrix(c(2,3,5,2,3,7,8,4,5),3,3))
}
