\name{mmds.gof}
\alias{mmds.gof}
\title{Goodness of fit for mixture model detection functions}
\description{
  Goodness of fit testing for detection for mixture model
  detection functions.
}
\section{Details}{
  Two goodness of fit tests are provided: the Cramer-von
  Mises and the Kolmogorov-Smirnov. Both are implemented as
  in Buckland et al. (2004).

  Print methods are provided, so accessing the \code{ks}
  and \code{cvm} elements of a \code{\link{ds.mixture}}
  object will give suitable summaries.

  David L. Miller
}

